refactor(git_custom_env): separate git custom environment
